+++ src/App.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import { HomePage } from "./film/movieList/components/movieList/MovieLsit";
-import { MovieInfo } from "./film/movieList/components/currentMovieInfo/MovieInfo";
-import Footer from "./film/movieList/components/footer";
+import { MovieList } from "./movieList/components/MovieList/MovieList"
+import { CurrentMovieInfo } from "./movieList/components/CurrentMovieInfo/CurrentMovieInfo";
+import Footer from "./movieList/components/Footer";
 
 import "./App.css";
 
@@ -11,10 +11,10 @@ function App() {
     <div className="App">
       <Router>
         <Routes>
-          <Route path="/" element={<HomePage />} />
+          <Route path="/" element={<MovieList />} />
         </Routes>
         <Routes>
-          <Route path="/movieInfo" element={<MovieInfo />} />
+          <Route path="/movieInfo" element={<CurrentMovieInfo />} />
         </Routes>
       </Router>
       <Footer />
