+++ src/app/store.js
@@ -1,5 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
-import movieReducer from "../film/movieList/redux/MovieListSlice";
+import movieReducer from "../movieList/redux/MovieListSlice";
 
 export const store = configureStore({
   reducer: {
