--- src/features/todolist/todolist.css
@@ -1,10 +1,0 @@
-.container__todo {
-  display: flex;
-  margin-top: 20px;
-  height: 100%;
-  flex-direction: column;
-}
-
-.list-group-item-success {
-  text-decoration: line-through;
-}
