--- src/film/movieList/components/movieList/notFound/NotFound.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-import "./notFound.css";
-
-export const NotFound = () => {
-  return <div className="block-not_movie">NO MOVIE FOUND</div>;
-};
