+++ src/movieList/components/Footer/footer.css
