+++ src/movieList/components/Footer/index.js
