+++ src/movieList/components/PageNotFound/PageNotFound.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+import "./pageNotFound.css";
+
+export const PageNotFound = () => {
+  return <div className="block-not_movie">NO MOVIE FOUND</div>;
+};
